+++ AtlasUI/AtlasUI/CheckoutSummaryViewController.swift
@@ -2,30 +2,23 @@
 //  Copyright © 2016 Zalando SE. All rights reserved.
 //
 
+import UIKit
 import AtlasSDK
 
-final class CheckoutSummaryViewController: UIViewController, CheckoutProviderType {
+class CheckoutSummaryViewController: UIViewController, CheckoutProviderType {
 
-    internal let productImageView = UIImageView()
-    internal let productNameLabel = UILabel()
-    internal let purchasedObjectSummaryLabel = UILabel()
-    internal let termsAndConditionsButton = UIButton()
-    internal let paymentSummaryTableview = UITableView()
-    internal let stackView: UIStackView = UIStackView()
-    internal let buyButton = UIButton()
-    internal let connectToZalandoButton = UIButton()
-
-    internal let shippingPrice: Float = 0
-    internal let shippingView = CheckoutSummaryRow()
-    internal let paymentMethodView = CheckoutSummaryRow()
-
-    internal private(set) var checkoutViewModel: CheckoutViewModel {
+    internal var checkout: AtlasCheckout
+    internal var checkoutViewModel: CheckoutViewModel
+    internal var viewState: CheckoutViewState = .NotLoggedIn {
         didSet {
-            updateData()
+//            setupSubmitButton()
+//            setupNavigationBar()
+//            refreshView()
         }
     }
-
-    internal var checkout: AtlasCheckout!
+    lazy private var actionsHandler: CheckoutSummaryActionsHandler = {
+        CheckoutSummaryActionsHandler(viewController: self)
+    }()
 
     init(checkout: AtlasCheckout, checkoutViewModel: CheckoutViewModel) {
         self.checkout = checkout
@@ -33,187 +26,157 @@ final class CheckoutSummaryViewController: UIViewController, CheckoutProviderTyp
 
         super.init(nibName: nil, bundle: nil)
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.setupViews()
-    }
+//    @IBOutlet private weak var articleImageView: UIImageView!
+//    @IBOutlet private weak var brandNameLabel: UILabel!
+//    @IBOutlet private weak var articleNameLabel: UILabel!
+//    @IBOutlet private weak var unitSizeLabel: UILabel!
+//    @IBOutlet private weak var shippingAddressTitleLabel: UILabel!
+//    @IBOutlet private weak var shippingAddressValueLabel: UILabel!
+//    @IBOutlet private weak var billingAddressTitleLabel: UILabel!
+//    @IBOutlet private weak var billingAddressValueLabel: UILabel!
+//    @IBOutlet private weak var paymentTitleLabel: UILabel!
+//    @IBOutlet private weak var paymentValueLabel: UILabel!
+//    @IBOutlet private weak var shippingTitleLabel: UILabel!
+//    @IBOutlet private weak var shippingPriceLabel: UILabel!
+//    @IBOutlet private weak var totalTitleLabel: UILabel!
+//    @IBOutlet private weak var totalPriceLabel: UILabel!
+//
+//    @IBOutlet private weak var footerLabel: UILabel!
+//    @IBOutlet private weak var submitButton: UIButton!
+//    @IBOutlet private weak var loaderView: UIView!
+//
+//    @IBOutlet private var arrowImageViews: [UIImageView] = []
+//    @IBOutlet private weak var priceStackView: UIStackView!
+//    @IBOutlet private weak var stackView: UIStackView! {
+//        didSet {
+//            stackView.subviews.flatMap { $0 as? UIStackView }.forEach {
+//                $0.layoutMargins = UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15)
+//                $0.layoutMarginsRelativeArrangement = true
+//            }
+//        }
+//    }
 
-    private func setupViews() {
-        Async.main {
-            self.view.removeAllSubviews()
-            self.stackView.removeAllSubviews()
-            self.title = self.loc("Summary")
-            self.view.backgroundColor = UIColor.clearColor()
-            self.view.opaque = false
-            self.setupNavBar()
-            self.setupBlurView()
-            self.setupProductImageView()
-            self.setupViewLabels()
-            self.setupStackView()
-            self.setupTermsAndConditionsButton()
-            self.setupButtons()
-            self.setupShippingView()
-            self.setupPaymentMethodView()
-            self.updateData()
-            CheckoutSummaryStyler(checkoutSummaryViewController: self).stylize()
-        }
-    }
+//    internal static func instantiateFromStoryBoard(checkout: AtlasCheckout,
+//                                                   checkoutViewModel: CheckoutViewModel) -> CheckoutSummaryStoryboardViewController? {
+//        let storyboard = UIStoryboard(name: "CheckoutSummaryStoryboard",
+//                                      bundle: NSBundle(forClass: CheckoutSummaryStoryboardViewController.self))
+//        guard let checkoutSummaryViewController = storyboard.instantiateInitialViewController()
+//            as? CheckoutSummaryStoryboardViewController else { return nil }
+//
+//        checkoutSummaryViewController.checkout = checkout
+//        checkoutSummaryViewController.checkoutViewModel = checkoutViewModel
+//        return checkoutSummaryViewController
+//    }
 
-    @objc internal func buyButtonTapped(sender: UIButton) {
-        let paymentProcessingViewController = PaymentProcessingViewController(checkout: checkout, checkoutViewModel: self.checkoutViewModel)
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//
+//        setupViewState()
+//    }
 
-        self.showViewController(paymentProcessingViewController, sender: self)
-    }
+//    internal func showLoader() {
+//        loaderView.hidden = false
+//    }
+//
+//    internal func hideLoader() {
+//        loaderView.hidden = true
+//    }
 
-    @objc internal func connectToZalandoButtonTapped(sender: UIButton) {
-        connectToZalando()
-    }
+//    dynamic private func cancelCheckoutTapped() {
+//        dismissView()
+//    }
+//
+//    private func dismissView() {
+//        dismissViewControllerAnimated(true, completion: nil)
+//    }
+//
+//    @IBAction private func submitButtonTapped() {
+//        switch viewState {
+//        case .NotLoggedIn: actionsHandler.loadCustomerData()
+//        case .LoggedIn: actionsHandler.handleBuyAction()
+//        case .OrderPlaced: dismissView()
+//        }
+//    }
+//
+//    @IBAction private func shippingAddressTapped() {
+//        guard viewState.showDetailArrow else { return }
+//
+//        userMessage.notImplemented()
+//    }
+//
+//    @IBAction private func billingAddressTapped() {
+//        guard viewState.showDetailArrow else { return }
+//
+//        userMessage.notImplemented()
+//    }
+//
+//    @IBAction private func paymentAddressTapped() {
+//        guard viewState.showDetailArrow else { return }
+//
+//        actionsHandler.showPaymentSelectionScreen()
+//    }
 
-    @objc private func cancelCheckoutTapped(sender: UIBarButtonItem) {
-        self.dismissViewControllerAnimated(true, completion: nil)
-    }
+}
 
-    private func showLoadingView() {
-        let indicator: UIActivityIndicatorView = UIActivityIndicatorView(activityIndicatorStyle: .WhiteLarge)
-        let backgroundView = UIView(frame: CGRect(x: 0, y: 0, width: self.view.bounds.size.width, height: self.view.bounds.size.height))
-        backgroundView.backgroundColor = .grayColor()
-        backgroundView.alpha = 0.2
-        indicator.center = CGPoint(x: self.view.bounds.size.width / 2, y: (self.view.bounds.size.height) / 2)
-        indicator.color = .blackColor()
-        Async.main {
-            indicator.startAnimating()
-            self.view.addSubview(backgroundView)
-            self.view.addSubview(indicator)
-        }
-    }
+extension CheckoutSummaryViewController {
 
-    private func setupNavBar() {
-        self.navigationItem.backBarButtonItem = UIBarButtonItem(title: "", style: .Plain, target: nil, action: nil)
-
-        if checkoutViewModel.article.hasSingleUnit {
-            self.navigationItem.setHidesBackButton(true, animated: false)
-        }
-
-        let cancelButton = UIBarButtonItem(title: loc("Cancel"), style: UIBarButtonItemStyle.Plain,
-            target: self, action: #selector(CheckoutSummaryViewController.cancelCheckoutTapped(_:)))
-
-        navigationItem.rightBarButtonItem = cancelButton
-    }
-
-    private func connectToZalando() {
-        loadCustomerData()
-    }
-
-    private func loadCustomerData() {
-        checkout.client.customer { result in
-            Async.main {
-                switch result {
-                case .failure(let error):
-                    self.userMessage.show(error: error)
-
-                case .success(let customer):
-                    self.showLoadingView()
-                    self.generateCheckoutAndRefreshViews(customer)
-                }
-            }
-        }
-    }
-
-    private func generateCheckoutAndRefreshViews(customer: Customer) {
-        checkout.createCheckout(withArticle: checkoutViewModel.article, selectedUnitIndex: checkoutViewModel.selectedUnitIndex) { result in
-            switch result {
-            case .failure(let error):
-                self.dismissViewControllerAnimated(true) {
-                    self.userMessage.show(error: error)
-                }
-            case .success(var checkout):
-                checkout.customer = customer
-                self.checkoutViewModel = checkout
-                self.setupViews()
-            }
-        }
-    }
-
-    private func setupProductImageView() {
-        view.addSubview(productImageView)
-    }
-
-    private func setupViewLabels() {
-        view.addSubview(productNameLabel)
-        view.addSubview(purchasedObjectSummaryLabel)
-    }
-
-    private func setupTermsAndConditionsButton() {
-        view.addSubview(termsAndConditionsButton)
-    }
-
-    private func setupButtons() {
-        view.addSubview(connectToZalandoButton)
-        connectToZalandoButton.hidden = true
-        connectToZalandoButton.setTitle(loc("Zalando.Connect"), forState: .Normal)
-        connectToZalandoButton.addTarget(self, action: #selector(CheckoutSummaryViewController.connectToZalandoButtonTapped(_:)),
-            forControlEvents: .TouchUpInside)
-        view.addSubview(buyButton)
-        buyButton.hidden = true
-        buyButton.setTitle(loc("Buy Now"), forState: .Normal)
-        buyButton.addTarget(self, action: #selector(CheckoutSummaryViewController.buyButtonTapped(_:)),
-            forControlEvents: .TouchUpInside)
-    }
-
-    private func setupStackView() {
-        view.addSubview(self.stackView)
-    }
-
-    private func setupBlurView() {
-        let blurEffect = UIBlurEffect(style: .ExtraLight)
-        let blurEffectView = UIVisualEffectView(effect: blurEffect)
-        view.addSubview(blurEffectView)
-
-        blurEffectView.frame = self.view.bounds
-        blurEffectView.frame.makeIntegralInPlace()
-
-        let vibrancy = UIVibrancyEffect(forBlurEffect: blurEffect)
-        let vibrancyView = UIVisualEffectView(effect: vibrancy)
-        vibrancyView.frame = blurEffectView.bounds
-        vibrancyView.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
-        let extraLightVibrancyView = vibrancyView
-        blurEffectView.contentView.addSubview(extraLightVibrancyView)
-    }
-
-    private func setupShippingView() {
-        shippingView.titleTextLabel.text = loc("Shipping")
-        shippingView.detailTextLabel.text = loc("No Shipping Address")
-        stackView.addArrangedSubview(shippingView)
-    }
-
-    private func setupPaymentMethodView() {
-        stackView.addArrangedSubview(paymentMethodView)
-
-        paymentMethodView.tapAction = {
-            guard let paymentURL = self.checkoutViewModel.checkout?.payment.selectionPageUrl else { return }
-            let paymentSelectionViewController = PaymentSelectionViewController(paymentSelectionURL: paymentURL)
-            paymentSelectionViewController.paymentCompletion = { _ in
-                self.loadCustomerData()
-            }
-            self.showViewController(paymentSelectionViewController, sender: self)
-        }
-    }
-
-    private func updateData() {
-        paymentMethodView.titleTextLabel.text = loc("Payment")
-        paymentMethodView.detailTextLabel.text = self.checkoutViewModel.paymentMethodText ?? loc("No Payment Method")
-
-        productNameLabel.text = checkoutViewModel.article.brand.name
-        purchasedObjectSummaryLabel.text = checkoutViewModel.article.name
-
-        productImageView.setImage(fromUrl: checkoutViewModel.article.thumbnailUrl)
-
-        shippingView.detailTextLabel.text = self.checkoutViewModel.shippingAddress(localizedWith: self.checkout)
-    }
+//    private func setupViewState() {
+//        if Atlas.isUserLoggedIn() {
+//            viewState = .LoggedIn
+//        } else {
+//            viewState = .NotLoggedIn
+//        }
+//    }
+//
+//    private func setupSubmitButton() {
+//        submitButton.setTitle(loc(viewState.submitButtonTitleLocalizedKey), forState: .Normal)
+//        submitButton.backgroundColor = viewState.submitButtonBackgroundColor
+//    }
+//
+//    private func setupNavigationBar() {
+//        title = loc(viewState.navigationBarTitleLocalizedKey)
+//
+//        let hasSingleUnit = checkoutViewModel.article.hasSingleUnit
+//        navigationItem.setHidesBackButton(viewState.hideBackButton(hasSingleUnit), animated: false)
+//
+//        if viewState.showCancelButton {
+//            let button = UIBarButtonItem(title: loc("Cancel"),
+//                                         style: .Plain,
+//                                         target: self,
+//                                         action: #selector(CheckoutSummaryStoryboardViewController.cancelCheckoutTapped))
+//            navigationItem.rightBarButtonItem = button
+//        } else {
+//            navigationItem.rightBarButtonItem = nil
+//        }
+//    }
+//
+//    private func refreshView() {
+//        hideLoader()
+//
+//        articleImageView.setImage(fromUrl: checkoutViewModel.article.thumbnailUrl)
+//        brandNameLabel.text = checkoutViewModel.article.brand.name
+//        articleNameLabel.text = checkoutViewModel.article.name
+//        unitSizeLabel.text = loc("Size: %@", checkoutViewModel.selectedUnit.size)
+//        shippingAddressTitleLabel.text = loc("Address.Shipping")
+//        shippingAddressValueLabel.text = checkoutViewModel.shippingAddress(localizedWith: self).trimmed
+//        billingAddressTitleLabel.text = loc("Address.Billing")
+//        billingAddressValueLabel.text = checkoutViewModel.billingAddress(localizedWith: self).trimmed
+//        paymentTitleLabel.text = loc("Payment")
+//        paymentValueLabel.text = checkoutViewModel.paymentMethodText
+//        shippingTitleLabel.text = loc("Shipping")
+//        shippingPriceLabel.text = localizer.fmtPrice(checkoutViewModel.shippingPriceValue)
+//        totalTitleLabel.text = loc("Total")
+//        totalPriceLabel.text = localizer.fmtPrice(checkoutViewModel.totalPriceValue)
+//        footerLabel.text = loc("CheckoutSummaryViewController.terms")
+//
+//        priceStackView.hidden = !viewState.showPrice
+//        footerLabel.hidden = !viewState.showFooterLabel
+//        arrowImageViews.forEach { $0.hidden = !viewState.showDetailArrow }
+//    }
 
 }
