+++ AtlasUI/AtlasUI/SizeListSelectionViewController.swift
@@ -8,7 +8,7 @@ import AtlasSDK
 final class SizeListSelectionViewController: UITableViewController, CheckoutProviderType {
 
     private let article: Article
-    internal let checkout: AtlasCheckout!
+    internal let checkout: AtlasCheckout
 
     init(checkout: AtlasCheckout, article: Article) {
         self.checkout = checkout
@@ -90,9 +90,7 @@ final class SizeListSelectionViewController: UITableViewController, CheckoutProv
     }
 
     private func displayCheckoutSummaryViewController(checkoutViewModel: CheckoutViewModel) {
-        if let checkoutSummaryVC = CheckoutSummaryStoryboardViewController
-            .instantiateFromStoryBoard(checkout, checkoutViewModel: checkoutViewModel) {
-            self.showViewController(checkoutSummaryVC, sender: self)
-        }
+        let checkoutSummaryVC = CheckoutSummaryViewController(checkout: checkout, checkoutViewModel: checkoutViewModel)
+        showViewController(checkoutSummaryVC, sender: self)
     }
 }
