+++ AtlasUI/AtlasUI/SizeSelectionViewController.swift
@@ -8,7 +8,7 @@ import AtlasSDK
 final class SizeSelectionViewController: UIViewController, CheckoutProviderType {
 
     private let sku: String
-    internal let checkout: AtlasCheckout!
+    internal let checkout: AtlasCheckout
 
     init(checkout: AtlasCheckout, sku: String) {
         self.checkout = checkout
@@ -70,11 +70,10 @@ final class SizeSelectionViewController: UIViewController, CheckoutProviderType 
     }
 
     private func displayCheckoutSummaryViewController(checkoutViewModel: CheckoutViewModel) {
-        if let checkoutSummaryVC = CheckoutSummaryStoryboardViewController
-            .instantiateFromStoryBoard(checkout, checkoutViewModel: checkoutViewModel) {
-            UIView.performWithoutAnimation {
-                self.showViewController(checkoutSummaryVC, sender: self)
-            }
+        let checkoutSummaryVC = CheckoutSummaryViewController(checkout: checkout, checkoutViewModel: checkoutViewModel)
+
+        UIView.performWithoutAnimation {
+            self.showViewController(checkoutSummaryVC, sender: self)
         }
     }
